+++ ha-integration/custom_components/ha_tv_pip/const.py
@@ -13,3 +13,4 @@ CONF_VERSION = "version"
 
 DEFAULT_PORT = 8765
 DISCOVERY_SERVICE_TYPE = "_ha-tv-pip._tcp.local."
+SERVICE_SHOW_CAMERA = "show_camera"